v1.51 check_emergency_reset, default break character is ctrl+z 3 times, example Arduino sketch
Added function from mungewell - check_emergency_reset. This has improved testing of the RX pin.	There was potential to get a false baud reset. There is still a chance but it's now much less likely.

If OpenLog is attached to a Arduino, during bootloading of the Arduino, ctrl+z will most likely be sent	to the Arduino from the computer. This character will cause OpenLog to drop to command mode - probably not what we want. So I added user selectable character (ctrl+x or '$' for example) and I added user selectable number of escape characters to look for in a row (example is 1 or 2 or 3, '$$$' is a common escape sequence). The default is now ctrl+z sent 3 times in a row.

Added an example Arduino sketch (from ScottH) to GitHub so that people can easily see how to send characters to	OpenLog. Not much to it, but it does allow us to test large amounts of text thrown at OpenLog at 57600bps.