v2.51 Improved command prompt parsing (now ignores '\n')
We now ignore the \n character when parsing characters from the command prompt. This makes it easier to create code on a microcontroller that correctly controls OpenLog. Previously, println or sprintf commands were adding a \n to the end of the string that would confuse OpenLog. One way around this if you have previous versions is this
 sprintf(buff, "new blah.txt\r");
 Serial.print(buff); //No println, use \r in string instead

Also - added a CommandTest sketch to demonstrate how you can control OpenLog from a microcontroller / automate the command prompt