v3.13 Lots of RAM management. Works well.
Receive buffer returned to its glory of being around 800 bytes. This is
very good news. Changed the string compares from strings in RAM (strcmp)
to strings located in flash memory(strcmp_P). Removal of strncmp(). Once
migrated to PSTR, it makes it worse. Issue 106 was the key to all this.