Better handling of max time before sync.
There is a case where the user could be throwing non-stop data at the
unit. The log won't be recorded because it never goes idle. This version
adds/makes better the check of max time for the un-recorded buffer. Also
moved from sprintf_P to strcpy_P for known file names.